Create utility module for test_broadcasting.py
